requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
motor>=3.3.0
python-dotenv>=1.0.0
langchain-openai>=0.1.0
langgraph>=0.2.0
langsmith>=0.1.0
langchain-core>=0.2.0
apscheduler>=3.10.0
bcrypt>=4.1.0
pydantic[email]>=2.5.0
reportlab>=4.0.0
resend>=0.8.0
httpx>=0.25.0
openai>=1.6.0
pandas>=2.1.0
openpyxl>=3.1.0
PyJWT>=2.8.0
boto3>=1.34.0
